[EU-78] Rename "components" dir to "common", replace here db
