+++ apps/main.py
@@ -1,11 +1,11 @@
 from apps.authorization.handlers import authorization_router
-from apps.components.exception_handlers import backend_exception_handler, http_exception_handler
-from apps.components.exceptions import BackendException
+from apps.common.exception_handlers import backend_exception_handler, http_exception_handler
+from apps.common.exceptions import BackendException
 from apps.educational_institutions.handlers import educational_institutions_router
 from apps.hostel.handlers import hostel_router
 from apps.services.handlers import services_router
 from apps.users.handlers import users_router
-from apps.core.db import database
+from apps.common.db import database
 from settings import Settings
 from tags_metadata import metadata
